modularização solidificada, ainda precisa trabalho
